Fix for Issue 58 - Some methods in Player don't pay attention to passed-in Synthesizer.
Created an instance variable to contain the Synthesizer. The variable is set if Player(Synthesizer) is constructed. The close() method pays attention to whether this variable is set, and if so, attempts to close the provided synthesizer.

There had already been a second static allNotesOff() method that took its own Synthesizer.

git-svn-id: http://jfugue.googlecode.com/svn/trunk@191 4e7ddc10-1079-11df-a8dc-870042ab4e57